# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/approval_policies:
    get:
      tags:
        - ApprovalPolicies
      summary: List ApprovalPolicies
      description: Get a paginated list of ApprovalPolicies with optional filtering and sorting
      operationId: listApprovalPolicies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalPolicyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ApprovalPolicies
      summary: Create ApprovalPolicy
      description: Create a new ApprovalPolicy
      operationId: createApprovalPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApprovalPolicyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalPolicy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/approval_requests:
    get:
      tags:
        - ApprovalRequests
      summary: List ApprovalRequests
      description: Get a paginated list of ApprovalRequests with optional filtering and sorting
      operationId: listApprovalRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ApprovalRequests
      summary: Create ApprovalRequest
      description: Create a new ApprovalRequest
      operationId: createApprovalRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApprovalRequestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    ApprovalPolicy:
      type: object
      required:
        - id
        - company_id
        - resource_type
        - name
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        resource_type:
          type: string
        name:
          type: string
        is_active:
          type: boolean
        description:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateApprovalPolicyRequest:
      type: object
      required:
        - company_id
        - resource_type
        - name
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        resource_type:
          type: string
        name:
          type: string
        is_active:
          type: boolean
        description:
          nullable: true
          type: string
        metadata:
          type: object
...
    ApprovalPolicyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalPolicy'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ApprovalRequest:
      type: object
      required:
        - id
        - company_id
        - resource_type
        - resource_id
        - requested_by
        - status
        - priority
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        policy_id:
          nullable: true
          type: string
          format: uuid
        requested_by:
          type: string
          format: uuid
        status:
          type: string
        current_step:
          nullable: true
          type: integer
          format: int32
        priority:
          type: string
        submitted_at:
          nullable: true
          type: string
          format: date-time
        decided_at:
          nullable: true
          type: string
          format: date-time
        decided_by:
          nullable: true
          type: string
          format: uuid
        summary:
          nullable: true
          type: object
        metadata:
          type: object
...
    CreateApprovalRequestRequest:
      type: object
      required:
        - company_id
        - resource_type
        - resource_id
        - requested_by
        - status
        - priority
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        policy_id:
          nullable: true
          type: string
          format: uuid
        requested_by:
          type: string
          format: uuid
        status:
          type: string
        current_step:
          nullable: true
          type: integer
          format: int32
        priority:
          type: string
        submitted_at:
          nullable: true
          type: string
          format: date-time
        decided_at:
          nullable: true
          type: string
          format: date-time
        decided_by:
          nullable: true
          type: string
          format: uuid
        summary:
          nullable: true
          type: object
        metadata:
          type: object
...
    ApprovalRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalRequest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []